test(matrix-utils): reorganize tests and add asyncio fixtures
migrate matrix utils tests to better module separation

introduce matrix_config fixture and align tests with pytest

add async tests for display-name helpers in room utilities
